requirements.txt
langchain
langchain-community
langchain-openai
langchain-chroma
pypdf
python-dotenv
chromadb
tiktoken